Update README.md
Hi! I'd like to add my game to the list. It uses Butterchurn as a background for a brick breaker. The game relies on accelerometer to detect subtle body movements to control the paddle.

Idk if my game should really be mentioned in the README since the it is a bit commercial: has banner ads in menu, sells energyyy that promises to make u feel better, but does nothing (except removing ads). However  there is also an open source component planned - it is a demo of "mind-controlled" menu, that has butterchurn as a background too just because (Mode:System > > Mode:Console, feel free to donate btc if it is usable enough to reach my btc address in it).

If you're an iPhone owner and you'd like to see butterchurn in a game, you can reach visuals through "Mode: Basic > Mode: MilkyWave". It is not default mode because it is heavy on GPU energyyy since there is alpha-blending on top of butterchurn. You can also tap "watch" (top of the screen in menu) to see pure butter churn (without alpha blending, so performance is better) and even PictureInPicture. The game constantly tracks fps and skips presets with fps less than 15, it also occasionally reduces fps of the game view (on next preset to 30fps until game-ball hits an obstacle and 2fps in menus) to reduce alpha-blending overhead when GPU is overheated (it greatly reduces stutter on iPhone SE 2). I'm open to less naive optimizations, feel free to reach me if you have any ideas. The blending is currently implemented through glkView.opaque=false (with glClear on every frame). 

P.S. I am not an expert in Computer Graphics, my domain is functional programming (Scala/Haskell), DSLs for financial data processing and DSLs for financial cryptography.
You can reach me at dmytro.kondratiuk.89@gmail.com. 🇺🇦